teams/requirements-teams.txt
# Bot Framework
botbuilder-core>=4.14.0
botbuilder-schema>=4.14.0

# Core dependencies from main project
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langgraph>=0.0.20
python-dotenv>=1.0.0

# Document generation
python-docx>=1.1.0

# Database and storage
sqlite3
aiosqlite>=0.19.0

# Vector database (if using Milvus)
pymilvus>=2.3.0

# Async support
asyncio
aiofiles>=23.0.0

# Utilities
pathlib
typing-extensions>=4.0.0
pydantic>=2.0.0